requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
ca-certificates=2023.08.22=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=pypi_0
cycler=0.11.0=pypi_0
essentia-tensorflow=2.1b6.dev1110=pypi_0
fonttools=4.38.0=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnsl=2.0.0=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libzlib=1.2.11=h166bdaf_1014
matplotlib=3.5.3=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.21.6=pypi_0
openssl=3.0.12=h7f8727e_0
packaging=24.0=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=py37h06a4308_0
portaudio=19.6.0=h7b6447c_4
pycparser=2.21=pypi_0
pyparsing=3.1.2=pypi_0
pyqt5=5.15.10=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.13.0=pypi_0
python=3.7.12=hf930737_100_cpython
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pypi_0
sounddevice=0.4.6=pypi_0
soundfile=0.12.1=pypi_0
sqlite=3.38.2=hc218d9a_0
tk=8.6.11=h1ccaba5_0
typing-extensions=4.7.1=pypi_0
wheel=0.38.4=py37h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.11=h166bdaf_1014
